# api/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.CreatePaymentRequest:
    properties:
      amount:
        type: number
      payment_method:
        type: string
      payment_status:
        type: string
      reservation_id:
        type: string
    type: object
  genproto.CreateReservationRequest:
    properties:
      reservation_time:
        type: string
      restaurant_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.PaymentResponse:
    properties:
      amount:
        type: number
      id:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      reservation_id:
        type: string
    type: object
  genproto.PaymentsResponse:
    properties:
      payments:
        items:
          $ref: '#/definitions/genproto.PaymentResponse'
        type: array
    type: object
  genproto.ReservationResponse:
    properties:
      reservation_time:
        type: string
      restaurant_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  genproto.ReservationsResponse:
    properties:
      Reservations:
        items:
          $ref: '#/definitions/genproto.ReservationResponse'
        type: array
    type: object
...
  genproto.UpdatePaymentRequest:
    properties:
      Id:
        type: string
      amount:
        type: number
      payment_method:
        type: string
      payment_status:
        type: string
      reservation_id:
        type: string
    type: object
  genproto.UpdateReservationRequest:
    properties:
      Id:
        type: string
      reservation_time:
        type: string
      restaurant_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: APi service
  version: "1.0"
paths:
...
  /api/payment/create:
    post:
      consumes:
      - application/json
      description: Create a new payment
      parameters:
      - description: Create Payment
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genproto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Payment
      tags:
      - Payment
  /api/payment/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Payment
      tags:
      - Payment
  /api/payment/get_all:
    get:
      consumes:
      - application/json
      description: Get all payments
      parameters:
      - in: query
        name: amount
        type: number
      - in: query
        name: payment_method
        type: string
      - in: query
        name: payment_status
        type: string
      - in: query
        name: reservation_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PaymentsResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Payments
      tags:
      - Payment
  /api/payment/get_id/{id}:
    get:
      consumes:
      - application/json
      description: Get a payment by its ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Payment by ID
      tags:
      - Payment
  /api/payment/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Payment
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Payment
      tags:
      - Payment
  /api/reservation/create:
    post:
      consumes:
      - application/json
      description: Create a new reservation
      parameters:
      - description: Create Reservation
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Reservation
      tags:
      - Reservation
  /api/reservation/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Reservation
      tags:
      - Reservation
  /api/reservation/get_all:
    get:
      consumes:
      - application/json
      description: Get all reservations
      parameters:
      - in: query
        name: reservation_time
        type: string
      - in: query
        name: restaurant_id
        type: string
      - in: query
        name: status
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ReservationsResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Reservations
      tags:
      - Reservation
  /api/reservation/get_id/{id}:
    get:
      consumes:
      - application/json
      description: Get a reservation by its ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Reservation by ID
      tags:
      - Reservation
  /api/reservation/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Reservation
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Reservation
      tags:
      - Reservation
...
swagger: "2.0"